# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Learning Management System backend API for GCEK.",
        "title": "GCEK LMS API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "GCEK LMS Support",
            "email": "support@gceklms.example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/code/run": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Executes code in a sandboxed environment without submitting it as an assignment.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "coding-assignments"
                ],
                "summary": "Run code in sandbox",
                "parameters": [
                    {
                        "description": "Code execution payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RunCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RunCodeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/courses/{id}/coding-assignments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all coding assignments for a course.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "coding-assignments"
                ],
                "summary": "List coding assignments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.CodingAssignmentResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new coding assignment with test cases for a course. Requires Teacher or Admin role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "coding-assignments"
                ],
                "summary": "Create a coding assignment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Coding assignment creation payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateCodingAssignmentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CodingAssignmentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/courses/{id}/coding-assignments/{codingAssignmentId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns details of a specific coding assignment including test cases.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "coding-assignments"
                ],
                "summary": "Get coding assignment by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Coding Assignment ID (UUID)",
                        "name": "codingAssignmentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CodingAssignmentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an existing coding assignment. Requires Teacher or Admin role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "coding-assignments"
                ],
                "summary": "Update a coding assignment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Coding Assignment ID (UUID)",
                        "name": "codingAssignmentId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateCodingAssignmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CodingAssignmentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a coding assignment. Requires Teacher or Admin role.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "coding-assignments"
                ],
                "summary": "Delete a coding assignment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Coding Assignment ID (UUID)",
                        "name": "codingAssignmentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CodingAssignmentResponse": {
            "type": "object",
            "properties": {
                "course_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "max_score": {
                    "type": "integer"
                },
                "starter_code": {
                    "type": "string"
                },
                "test_cases": {
                    "description": "hidden test cases stripped for students",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TestCase"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "dto.CreateCodingAssignmentRequest": {
            "type": "object",
            "required": [
                "language",
                "max_score",
                "test_cases",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "due_date": {
                    "description": "ISO 8601 optional",
                    "type": "string"
                },
                "language": {
                    "type": "string",
                    "enum": [
                        "python",
                        "javascript"
                    ]
                },
                "max_score": {
                    "type": "integer"
                },
                "starter_code": {
                    "type": "string"
                },
                "test_cases": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TestCase"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "dto.RunCodeRequest": {
            "type": "object",
            "required": [
                "code",
                "language"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "input": {
                    "description": "arbitrary stdin for quick testing",
                    "type": "string"
                },
                "language": {
                    "type": "string",
                    "enum": [
                        "python",
                        "javascript"
                    ]
                }
            }
        },
        "dto.RunCodeResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "execution_time_ms": {
                    "type": "integer"
                },
                "output": {
                    "type": "string"
                },
                "stderr": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}